requirement.txt
awscli
boto3
bs4
click
datetime
docker-py
pyyaml
wget